# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Enterprise Doctor Verification System API
  description: Production-grade REST API specification for Practo-like Doctor KYC and Identity Verification system.
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Local Development Server
  - url: https://verification.practo-doctor.portal/api/v1
    description: Production Kubernetes Cluster
paths:
...
  /doctors/verify-otp:
    post:
      summary: Verify Mobile OTP
      operationId: verifyOTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [otp, purpose]
              properties:
                public_id: { type: string }
                otp: { type: string, example: "123456" }
                purpose: { type: string, example: "REGISTER" }
      responses:
        '200':
          description: Mobile verified, JWT access and refresh tokens returned.
...
  /doctors/documents:
    post:
      summary: Secure Multipart Document Upload
      operationId: uploadDocument
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, document_type]
              properties:
                file: { type: string, format: binary }
                document_type: { type: string, example: "REGISTRATION_CERTIFICATE" }
      responses:
        '201':
          description: Document uploaded securely, hash and version stored.
...